add: 긱 필터 테스트, EMA의 알파값 조정 코드드
